chore: rename hook
